backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
websockets==12.0
pyyaml==6.0.3
cryptography==41.0.7
requests==2.31.0
httpx==0.25.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
argon2-cffi==23.1.0
